refactor(comb): Split up module
